added sort by functionality, link to cor. author's orcid page.

+++ website/views.py
@@ -5,6 +5,7 @@ from django.shortcuts import render
 from dotenv import load_dotenv
 from bs4 import BeautifulSoup
 from .models import Record, Author
+from .utils.soup_helper import get_parsed_html
 
 
 def home(request):
@@ -15,7 +16,6 @@ def home(request):
     records = Record.objects.all()
     return render(request, 'search_results.html', {'records': records})'''
 
-
 load_dotenv()
 api_key = os.getenv('API_KEY')
 
@@ -23,7 +23,11 @@ api_key = os.getenv('API_KEY')
 def search(request):
     if request.method == 'POST':
         query = request.POST.get('query')
-        articles_number = int(request.POST.get('articles'))
+
+        if not query:
+            return render(request, 'home.html', {'error': 'Please enter a search query.'})
+
+        articles_number = int(request.POST.get('articles', 25))
         custom_articles_number = request.POST.get('custom_articles')
         Record.objects.all().delete()
         Author.objects.all().delete()
@@ -79,8 +83,12 @@ def search(request):
         for paper in papers:
             authors_list = paper['authors'].split(', ')
             authors_objs = []
+            pii = paper['pii']
             for author_name in authors_list:
-                author, created = Author.objects.get_or_create(name=author_name.strip())
+                orcid_id = None
+                if author_name == paper['corresponding_author']:
+                    orcid_id = search_corresponding_author_orcid_id(pii)
+                author, created = Author.objects.get_or_create(name=author_name.strip(), defaults={'orcid_id': orcid_id})
                 authors_objs.append(author)
             record = Record.objects.create(
                 title=paper['title'],
@@ -92,16 +100,17 @@ def search(request):
             record.authors.add(*authors_objs)
 
         records = Record.objects.all()
+        sort_by = request.POST.get('sort_by', 'relevance')
+        records = show_articles_by_date(records, sort_by)
+
         return render(request, 'search_results.html', {'records': records})
     else:
         return render(request, 'home.html', {})
 
 
 def search_corresponding_author_name(pii):
-    url = f"https://api.elsevier.com/content/article/pii/{pii}?apiKey={api_key}"
-    response = requests.get(url)
-    data = response.text
-    soup = BeautifulSoup(data, "lxml")
+    pii_url = f"https://api.elsevier.com/content/article/pii/{pii}?apiKey={api_key}"
+    soup = get_parsed_html(pii_url)
 
     author_tags = soup.find_all('ce:author')
 
@@ -119,5 +128,28 @@ def search_corresponding_author_name(pii):
         return "Author not found."
 
 
+def search_corresponding_author_orcid_id(pii):
+    pii_url = f"https://api.elsevier.com/content/article/pii/{pii}?apiKey={api_key}"
+    soup = get_parsed_html(pii_url)
+
+    author_tags = soup.find_all('ce:author')
+
+    for author_tag in author_tags:
+        if author_tag.find('ce:cross-ref', {'refid': "cor1"}) or author_tag.find('ce:cross-ref', {'refid': "cr0005"}):
+            orcid_id = author_tag.get('orcid')
+            if orcid_id:
+                return orcid_id
+            else:
+                return soup.find('ce:author').get('orcid')
+
+
 def search_corresponding_author_email(corresponding_author_name):
     return "Email not found."
+
+
+def show_articles_by_date(records, sort_by):
+    if sort_by == 'newest':
+        records = records.order_by('-date')
+    elif sort_by == 'oldest':
+        records = records.order_by('date')
+    return records
